Added changed tooltip to the bar charts and different redirect options
