+++ src/Charts/GroupedBarChart/GroupedBarChart.test.js
