+++ src/Charts/GroupedBarChart/index.js
@@ -1,0 +1,3 @@
+export { default as GroupedBarChart } from './GroupedBarChart';
+export { default as OrgsTooltip } from './OrgsTooltip';
+export { default as HostsTooltip } from './HostsTooltip';
